requirements.txt
click==8.1.2
colorama==0.4.4
Flask==2.1.1
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
mysql-connector==2.2.9
Werkzeug==2.1.1
